requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.3
email-validator==2.1.1
python-multipart==0.0.20
